Isolate disk sync logic behind an interface
